ajout du dubbing pour les fichiers

+++ back/api/dubbing_utils.py
@@ -19,51 +19,60 @@ if not ELEVENLABS_API_KEY:
 
 client = ElevenLabs(api_key=ELEVENLABS_API_KEY)
 
+def is_audio_file(file_name: str) -> bool:
+    audio_extensions = ['.mp3', '.wav', '.flac', '.aac', '.ogg']
+    return any(file_name.lower().endswith(ext) for ext in audio_extensions)
 
 def download_dubbed_file(dubbing_id: str, language_code: str) -> str:
+
+    headers = {"xi-api-key": f"{ELEVENLABS_API_KEY}"}
+    url_metadata = f"https://api.elevenlabs.io/v1/dubbing/{dubbing_id}"
+    response_metadata = requests.get(url_metadata, headers=headers)
+
+    is_audio_type = False
+
+    if response_metadata.status_code == 200:
+        metadata = response_metadata.json()
+        file_name = metadata.get("name", "")
+        is_audio_type = is_audio_file(file_name)
 
     dir_path = "data/"
     os.makedirs(dir_path, exist_ok=True)
 
-    url = f"https://api.elevenlabs.io/v1/dubbing/{dubbing_id}/audio/en"
+    url = f"https://api.elevenlabs.io/v1/dubbing/{dubbing_id}/audio/{language_code}"
 
-    headers = {"xi-api-key": f"{ELEVENLABS_API_KEY}"}
-
-    response = requests.request("GET", url, headers=headers)
+    response = requests.get(url, headers=headers)
 
     if response.status_code == 200:
 
         dir_path = f"data/{dubbing_id}"
         os.makedirs(dir_path, exist_ok=True)
 
-        mp4_file_path = f"{dir_path}/{language_code}.mp4"
-        with open(mp4_file_path, "wb") as file:
-            file.write(response.content)
+        if not is_audio_type :
+            mp4_file_path = f"{dir_path}/{dubbing_id}_{language_code}.mp4"
+            with open(mp4_file_path, "wb") as file:
+                file.write(response.content)
 
-        # Convert MP4 to MP3
-        mp3_file_path = f"{dir_path}/{language_code}.mp3"
-        audio = AudioSegment.from_file(mp4_file_path, format="mp4")
-        audio.export(mp3_file_path, format="mp3")
+            # Convert MP4 to MP3
+            mp3_file_path = f"{dir_path}/{dubbing_id}_{language_code}.mp3"
+            audio = AudioSegment.from_file(mp4_file_path, format="mp4")
+            audio.export(mp3_file_path, format="mp3")
 
-        # Supprimer le fichier MP4 après la conversion
-        os.remove(mp4_file_path)
-
+            # Supprimer le fichier MP4 après la conversion
+            os.remove(mp4_file_path)
+        else :
+            mp3_file_path = f"{dir_path}/{dubbing_id}_{language_code}.mp3"
+            with open(mp3_file_path, "wb") as file:
+                file.write(response.content)
+                   
         return mp3_file_path
+
     else:
         print(f"Failed to download the file: {response.status_code}, {response.text}")
         return ""
 
 
 def wait_for_dubbing_completion(dubbing_id: str) -> bool:
-    """
-    Waits for the dubbing process to complete by periodically checking the status.
-
-    Args:
-        dubbing_id (str): The dubbing project id.
-
-    Returns:
-        bool: True if the dubbing is successful, False otherwise.
-    """
     MAX_ATTEMPTS = 120
     CHECK_INTERVAL = 10  # In seconds
